pyproject.toml
[project]
name = "generalaiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio[oauth]>=5.26.0",
    "hydra-core>=1.3.2",
    "langfuse>=2.60.3",
    "loguru>=0.7.3",
    "numpy>=2.2.5",
    "openai>=1.76.0",
    "openinference-instrumentation-smolagents>=0.1.10",
    "opentelemetry-exporter-otlp>=1.32.1",
    "opentelemetry-sdk>=1.32.1",
    "pandas>=2.2.3",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "smolagents[openai,telemetry]>=1.14.0",
    "tqdm>=4.67.1",
]
